Actual fix that avoids recompiling Emacs C sources for dlopen with RTLD_GLOBAL.
Previous commit claims to fix this, but it actually didn't.

Ugly workaround to being required to load libjulia with RTLD_GLOBAL.
We load it first via the wrapper, which has a custom dlopen call
Then, emacs-ffi tries to re-load it via dtld, but it's already
been loaded with the RTLD_GLOBAL flag from the wrapper.
Note that ltld docs claim their dlopen shouldn't need RTLD_GLOBAL because
"back-tracing". This doesn't seem to be true for libjulia...